+++ glide-core/src/socket_listener.rs
@@ -363,6 +363,7 @@ fn get_command(request: &Command) -> Option<Cmd> {
         RequestType::DBSize => Some(cmd("DBSIZE")),
         RequestType::Brpop => Some(cmd("BRPOP")),
         RequestType::PfAdd => Some(cmd("PFADD")),
+        RequestType::PfCount => Some(cmd("PFCOUNT")),
     }
 }