+++ java/integTest/src/test/java/glide/TransactionTestUtilities.java
@@ -20,6 +20,7 @@ public class TransactionTestUtilities {
     private static final String key8 = "{key}" + UUID.randomUUID();
     // TODO rename after #160
     private static final String hllKey1 = "{key}:hllKey1-" + UUID.randomUUID();
+    private static final String hllKey2 = "{key}:hllKey2-" + UUID.randomUUID();
     private static final String value1 = UUID.randomUUID().toString();
     private static final String value2 = UUID.randomUUID().toString();
     private static final String value3 = UUID.randomUUID().toString();
@@ -102,6 +103,7 @@ public class TransactionTestUtilities {
         baseTransaction.echo("GLIDE");
 
         baseTransaction.pfadd(hllKey1, new String[] {"a", "b", "c"});
+        baseTransaction.pfcount(new String[] {hllKey1, hllKey2});
 
         return baseTransaction;
     }
@@ -162,6 +164,7 @@ public class TransactionTestUtilities {
             OK,
             "GLIDE", // echo
             1L, // pfadd(hllKey1, new String[] {"a", "b", "c"})
+            3L, // pfcount(new String[] { hllKey1, hllKey2 });
         };
     }
 }