+++ CHANGELOG.md
@@ -1,6 +1,7 @@
 #### Changes
 * Python: Added JSON.DEL JSON.FORGET commands  ([#1146](https://github.com/aws/glide-for-redis/pull/1146))
 * Core: Added `PFADD` command. ([#1221](https://github.com/aws/glide-for-redis/pull/1221))
+* Core: Added `PFCOUNT` command. ([#1222](https://github.com/aws/glide-for-redis/pull/1222))
 
 #### Fixes
 * Python: Fix typing error "‘type’ object is not subscriptable" ([#1203](https://github.com/aws/glide-for-redis/pull/1203))